Improve Shop Page.
